[Frontend] Moved CodeEditorWithSyntax to components/CodeEditorWithSyntax folder
