+++ frontend/src/components/CodeEditorWithSyntax/CodeEditorWithSyntax.tsx
